pyproject.toml
[tool.poetry]
name = "testiky"
version = "0.1.0"
description = ""
authors = ["Michal Janočko <michal@janocko.com>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.91.0"}
python-dotenv = "^0.21.1"
aiofiles = "^23.1.0"
pydantic = "^1.10.4"
email-validator = "^2.0.0.post2"
aiohttp = "^3.8.4"
asyncpg = "^0.28.0"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.19"
requests = "^2.31.0"
rich = "^13.4.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
